Allow overriding the maximum event count
The library handles up to ten events by default, as defined by the
preprocessor macro MAX_NUMBER_OF_EVENTS in Timer.h.  However,
sometimes that's just not enough.  To change this default one would
have to edit the header file, meaning the change would affect all
other programs using the library as well: that's not good.

So let's add conditional preprocessor directives to define the default
value in Timer.h only when a value hasn't already been defined, thus
allowing users to override the default without modifying the library's
source files.  Also change the name of the preprocessor macro by
adding the "TIMER_" prefix, because "MAX_NUMBER_OF_EVENTS" was just
too general.

Note that a simple "#define TIMER_MAX_NUMBER_OF_EVENTS 42" in the
source file before #including Timer.h is not sufficient, because that
macro is not visible when Timer.cpp is compiled.  The overriding macro
must be defined as a compiler option.  It's easy enough to put
-DTIMER_MAX_NUMBER_OF_EVENTS=42 into CXXFLAGS when building with
Arduino-Makefile, but don't know how (or whether) that could be done
with the Arduino IDE.  (Alternatively, we could simply pass the
desired number of events as an optional constructor parameter and
dynamically allocate the array of Events, but this library doesn't use
dynamic memory allocation, and we would probably want to keep it that
way.)