Simplify error checking in Timer::pulseImmediate()
Timer::pulseImmediate() is a bit more defensive than necessary, as it
checks that the event id returned from oscillate() falls withing the
allowed range.  It always does, unless oscillate() returns with a
negative error code, so drop the unnecessary upper bound check.

While at it, use the the good old '=' to assign the value of the local
'id' variable instead of using a constructor-looking construct.