Store the time an Event is to be executed next
Store the time an Event is to be executed next, not when it was last
executed.  This simplifies the condition in Event::update() a tad bit,
sparing a substraction on each call, even when the event's action
wasn't executed.  More importantly, it will also allow us to easily
implement methods to execute periodic events starting after a given
delay.