Add methods to create periodic events starting after a delay
Both Timer::every() and Timer::oscillate() create timer events that
start immediately and are executed for the first time after the given
time period is up.  However, it's sometimes necessary to create timers
that start after a specific delay different from the timer's period,
and these are cumbersome to do with the current Timer API.

So let's add overloaded variants of the every() and oscillate()
methods that accept an additional 'after' parameter specifying the
initial delay before the event is first executed.